counting good, still need to test with larger photos, improved log (now check if exist file or folder and if not it creates them)

+++ src/log.php
@@ -1,5 +1,7 @@
 <?php
 
+define("LOG_FOLDER","logs");
+define("LOG_FILE","log.txt");
 
 class Log{
 
@@ -7,7 +9,18 @@ class Log{
     public function WriteLog($msg)
     {
       $current_datetime=date("d.m.y H:i:s");
-      $file="./logs/log.txt";
+      $file="./".LOG_FOLDER."/".LOG_FILE;
+      if(!file_exists($file))
+      {
+         if(!file_exists("./".LOG_FOLDER))
+         {
+           mkdir("./".LOG_FOLDER);
+         }
+
+         fopen("./".LOG_FOLDER."/".LOG_FILE, 'w') or die();
+      
+      }
+     
       file_put_contents($file,$msg."   |".$current_datetime."  \n",FILE_APPEND| LOCK_EX);
